Do not use mmaped files, read segments via file API
The theory was: on HT CPU when we wait for IO on one thread, another thread runs. But with page cache we probably do not see page faults at all after the first run.

Same perf but no mmap, which is probably a win.

Processed in 00:00:03.2365494
Processed in 00:00:03.1375309
Processed in 00:00:03.2481944
Processed in 00:00:03.1867151